Added dotenv configuration options.
Updated index.js to use process.env.<VAR>
instead of hardcoding the configuration options
in.

Also moved the generation of new api keys to their own file.
