+++ .env.sample
@@ -1,0 +1,3 @@
+STRIPE_SECRET_KEY=sk_test-YOUR_SECRET_KEY
+STRIPE_PRICE_KEY=pk_YOUR_PRICE_KEY
+WEBSITE_URL=localhost:8080
